Feature/cmake cleanup (#381)
* Rename OnScreenGui module to ImGui
* Support multiple external module folders (closes #31)
* Cleaning up CMake files
  * Restructure application specification
  * Add parameter for library mode to all modules
  * Add functions to handle global variable state
* Misc/remove warnings (#383)
* Increase build timeout and do a clean rebuild every commit

+++ modules/newhorizons/util/labelparser.cpp
@@ -330,7 +330,7 @@ void LabelParser::createImage(Image& image, double startTime, double stopTime, s
     image.timeRange = { startTime , stopTime };
     ghoul_assert(image.timeRange.isDefined(), "Invalid time range!");
     image.path = pot;
-    for (int i = 0; i < instr.size(); i++){
+    for (size_t i = 0; i < instr.size(); i++){
         image.activeInstruments.push_back(instr[i]);
     }
     image.target = targ;
